requirements.txt
fastapi
uvicorn
langchain
openai
PyMuPDF
tiktoken
faiss-cpu
python-multipart
pydantic
